+++ src/baryCentric.h
@@ -2,14 +2,14 @@
 #include "types.h"
 
 struct BaryCentric {
-    BaryCentric(RenderVertex const &b, RenderVertex const &c, RenderVertex const &a, Vec3d const &vertexDist)
+    BaryCentric(RenderVertex const &b, RenderVertex const &c, RenderVertex const &a, Vec3f const &vertexDist)
         : _vertexDist(vertexDist) {
         _a = a.imageCoord;
         v0 = b.imageCoord - a.imageCoord;
         v1 = c.imageCoord - a.imageCoord;
-        d00 = v0.cast<double>().dot(v0.cast<double>());
-        d01 = v0.cast<double>().dot(v1.cast<double>());
-        d11 = v1.cast<double>().dot(v1.cast<double>());
+        d00 = v0.cast<float>().dot(v0.cast<float>());
+        d01 = v0.cast<float>().dot(v1.cast<float>());
+        d11 = v1.cast<float>().dot(v1.cast<float>());
         denom = d00 * d11 - d01 * d01;
 
         texCoords.col(0) = a.textureCoord;
@@ -17,20 +17,20 @@ struct BaryCentric {
         texCoords.col(2) = c.textureCoord;
     }
 
-    inline std::pair<Vec2d, double> ConvertImageToTextureDist(Vec2i const &p) const {
+    inline std::pair<Vec2f, float> ConvertImageToTextureDist(Vec2i const &p) const {
         Vec2i v2 = p - _a;
-        double d20 = v2.cast<double>().dot(v0.cast<double>());
-        double d21 = v2.cast<double>().dot(v1.cast<double>());
+        float d20 = v2.cast<float>().dot(v0.cast<float>());
+        float d21 = v2.cast<float>().dot(v1.cast<float>());
 
-        Vec3d bWeights;
+        Vec3f bWeights;
         bWeights[0] = (d11 * d20 - d01 * d21) / denom;
         bWeights[1] = (d00 * d21 - d01 * d20) / denom;
         bWeights[2] = 1.0f - bWeights[0] - bWeights[1];
         return {texCoords * bWeights, bWeights.dot(_vertexDist)};
     }
     Vec2i _a;
-    double d00, d01, d11, denom;
-    Vec3d _vertexDist; // dist from camera
+    float d00, d01, d11, denom;
+    Vec3f _vertexDist; // dist from camera
     Vec2i v0, v1;
-    Eigen::Matrix<double, 2, 3> texCoords; // mapping from vertices to texture coordinates
+    Eigen::Matrix<float, 2, 3> texCoords; // mapping from vertices to texture coordinates
 };
