+++ src/types.h
@@ -3,14 +3,13 @@
 
 using Vec2i = Eigen::Vector2i;
 using Vec3i = Eigen::Vector3i;
-using Vec2d = Eigen::Vector2d;
-using Vec3d = Eigen::Vector3d;
-using Vec4d = Eigen::Vector4d;
-using Mat3d = Eigen::Matrix3d;
+using Vec2f = Eigen::Vector2f;
+using Vec3f = Eigen::Vector3f;
+using Vec4f = Eigen::Vector4f;
+using Mat3f = Eigen::Matrix3f;
 
 struct RenderVertex {
-    RenderVertex(Vec2i const &im, Vec2d const &tx) : imageCoord(im), textureCoord(tx) {}
+    RenderVertex(Vec2i const &im, Vec2f const &tx) : imageCoord(im), textureCoord(tx) {}
     Vec2i imageCoord;
-    Vec2d textureCoord;
+    Vec2f textureCoord;
 };
-
